Introduce Department model and migration
Laravel uses an ORM (Object Relational Mapper) called 'Eloquent' which
makes it easy to interact with relational data from PHP objects. This
commit introduces an Eloquent model for department - which extends
'Model'. This might look empty but has a huge amount of functionality
inherited from 'Model'. Eloquent implements the ActiveRecord pattern.

The migration is like version control for a database - allowing a team
to share updates to a database schema.

Key things to show:

- Artisan commands for model and migration creation
- Artisan migration commands