Introduce browser tests
Adds a couple of simple automated browser tests which make use of
the capabilities provided by Laravel Dusk. Note: these will spin up
a Chrome instance when run so you will be able to see the interactions
in real-time.

Laravel Dusk provides several methods for interacting with a page. See
[the docs](https://laravel.com/docs/5.4/dusk#interacting-with-elements)
for more.